update: js helper
