+++ opts/global.go
@@ -1,7 +1,6 @@
 package opts
 
 import (
-	"fmt"
 	"math/rand"
 	"reflect"
 	"regexp"
@@ -291,37 +290,6 @@ type Indicator struct {
 
 var newlineTabPat = regexp.MustCompile(`\n|\t`)
 var commentPat = regexp.MustCompile(`(//.*)\n`)
-
-const funcMarker = "__f__"
-
-type JSFunctions struct {
-	Fns []string
-}
-
-// AddJSFuncs adds a new JS function.
-func (f *JSFunctions) AddJSFuncs(fn ...string) {
-	for i := 0; i < len(fn); i++ {
-		f.Fns = append(f.Fns, newlineTabPat.ReplaceAllString(fn[i], ""))
-	}
-}
-
-// FuncOpts returns a string suitable for options expecting JavaScript code.
-func FuncOpts(fn string) string {
-	return replaceJsFuncs(fn)
-}
-
-// FuncStripCommentsOpts returns a string suitable for options expecting JavaScript code,
-// stripping '//' comments.
-func FuncStripCommentsOpts(fn string) string {
-	fn = commentPat.ReplaceAllString(fn, "")
-	return replaceJsFuncs(fn)
-}
-
-// replace and clear up js functions string
-func replaceJsFuncs(fn string) string {
-	fn = newlineTabPat.ReplaceAllString(fn, "")
-	return fmt.Sprintf("%s%s%s", funcMarker, fn, funcMarker)
-}
 
 type Colors []string
